formatting and type hinting

+++ ext/voice.py
@@ -1,5 +1,6 @@
 import asyncio
 
+import discord
 from discord.ext import commands
 
 from utils import AloneBot
@@ -10,8 +11,14 @@ class Voice(commands.Cog):
         self.bot = bot
 
     @commands.Cog.listener()
-    async def on_voice_join(self, member, state):
+    async def on_voice_join(self, member: discord.Member, state: discord.VoiceState):
+        assert state.channel
         vc = self.bot.guild_configs.get(state.channel.guild.id, {}).get("voice_channel", None)
+        category_channel_id = self.bot.guild_configs.get(state.channel.guild.id, {}).get("voice_category", None)
+
+        if not category_channel_id:
+            return
+
         if not vc or state.channel.id != vc:
             return
 
@@ -24,11 +31,10 @@ class Voice(commands.Cog):
 
         new_vc = await member.guild.create_voice_channel(
             name=member.display_name,
-            category=member.guild.get_channel(
-                self.bot.guild_config.get(state.channel.guild.id, None).get("voice_category", None)
-            ),
+            category=member.guild.get_channel(category_channel_id),  # type: ignore
             reason="Made by the personal voice chat module",
         )
+
         await member.move_to(channel=new_vc)
         self.bot.guild_configs[member.guild.id].setdefault("community_voice_channels", {})[new_vc.id] = member.id
         await self.bot.db.execute(
@@ -40,7 +46,9 @@ class Voice(commands.Cog):
         await member.send("Welcome to your own voice chat! Here, you lay the rules. your house, your magic. Have fun!")
 
     @commands.Cog.listener()
-    async def on_voice_leave(self, member, state):
+    async def on_voice_leave(self, member: discord.Member, state: discord.VoiceState):
+        assert state.channel
+
         vc = self.bot.guild_configs.get(member.guild.id, {}).get("community_voice_channels", {})
         if not vc or not state.channel.id in vc:
             return
@@ -48,19 +56,27 @@ class Voice(commands.Cog):
         if not state.channel.members:
             channel = state.channel
 
-            def channel_check(_, state):
+            def channel_check(_: discord.Member, state: discord.VoiceState) -> bool:
                 return state.channel == channel
 
             try:
-                owner = (self.bot.get_guild(member.guild.id)).get_member(vc.get(state.channel.id))
+                owner_id = vc.get(state.channel.id)
+                assert owner_id
+
+                owner = member.guild.get_member(owner_id)
+                assert owner
+
                 message = await owner.send(
                     "I will delete your private channel for inactivity in 5 minutes if it's not used!"
                 )
                 await self.bot.wait_for("voice_join", timeout=300, check=channel_check)
                 await message.delete()
+
             except asyncio.TimeoutError:
                 try:
-                    await channel.delete()
+                    if channel:
+                        await channel.delete()
+
                     await self.bot.db.execute("DELETE FROM voice WHERE channel_id = $1", state.channel.id)
                 except Exception:
                     pass
